feat(checkout-providers): moved checkout tests into separate folder
