+++ packages/falcon-front-kit/src/Checkout/__tests__/SetShippingMethod.test.tsx
@@ -7,28 +7,22 @@ import { SchemaLink } from 'apollo-link-schema';
 import { InMemoryCache } from 'apollo-cache-inmemory';
 import { ApolloClient } from 'apollo-client';
 import { ApolloProvider, MutationResult } from '@apollo/react-common';
-import { SetBillingAddressResponse } from '@deity/falcon-shop-data';
-import {
-  Checkout,
-  CheckoutRenderProps,
-  CheckoutProvider,
-  CheckoutOperationFunction,
-  SetBillingAddress,
-  CheckoutAddress
-} from '.';
+import { SetShippingMethodResponse } from '@deity/falcon-shop-data';
+import { ShippingMethod } from '@deity/falcon-shop-extension';
+import { Checkout, CheckoutRenderProps, CheckoutProvider, CheckoutOperationFunction, SetShippingMethod } from '..';
 
 const BaseSchema = readFileSync(require.resolve('@deity/falcon-server/schema.graphql'), 'utf8');
 const Schema = readFileSync(require.resolve('@deity/falcon-shop-extension/schema.graphql'), 'utf8');
 
-const sampleAddress: CheckoutAddress = {
-  id: 1,
-  firstname: 'foo',
-  lastname: 'bar',
-  city: 'Sample City',
-  postcode: '00000',
-  street: ['Sample Street 12'],
-  country: { id: 'NL', code: 'NL' },
-  telephone: '000000000'
+const sampleShippingMethod = {
+  carrierTitle: 'test carrier',
+  amount: 10,
+  carrierCode: 'test',
+  methodCode: 'test',
+  methodTitle: 'Test Carrier',
+  priceExclTax: 8,
+  priceInclTax: 10,
+  currency: 'EUR'
 };
 
 const createApolloClient = (resolvers: any) => {
@@ -50,15 +44,15 @@ const createApolloClient = (resolvers: any) => {
   });
 };
 
-describe('SetBillingAddress', () => {
-  describe('when setting billing address', () => {
+describe('SetShippingMethod', () => {
+  describe('when setting shipping method', () => {
     let wrapper: ReactWrapper<any, any> | null;
     let client: ApolloClient<any>;
 
     beforeEach(() => {
       client = createApolloClient({
         Mutation: {
-          setBillingAddress: () => true
+          setShippingMethod: () => true
         }
       });
     });
@@ -70,9 +64,9 @@ describe('SetBillingAddress', () => {
       }
     });
 
-    it('should properly set billing address on Checkout context', async () => {
-      let checkoutOperation: CheckoutOperationFunction<SetBillingAddressResponse, CheckoutAddress>;
-      let checkoutOperationResult: MutationResult<SetBillingAddressResponse>;
+    it('should properly set shipping method on Checkout context', async () => {
+      let checkoutOperation: CheckoutOperationFunction<SetShippingMethodResponse, ShippingMethod>;
+      let checkoutOperationResult: MutationResult<SetShippingMethodResponse>;
       let props: CheckoutRenderProps;
       let isLoading: boolean;
       wrapper = mount(
@@ -86,14 +80,14 @@ describe('SetBillingAddress', () => {
                 }
 
                 return (
-                  <SetBillingAddress>
+                  <SetShippingMethod>
                     {(operation, result) => {
                       checkoutOperation = operation;
                       checkoutOperationResult = result;
 
                       return <div />;
                     }}
-                  </SetBillingAddress>
+                  </SetShippingMethod>
                 );
               }}
             </Checkout>
@@ -102,14 +96,14 @@ describe('SetBillingAddress', () => {
       );
 
       expect(props.isLoading).toBeFalsy();
-      expect(props.values.billingAddress).toBeUndefined();
+      expect(props.values.shippingMethod).toBeUndefined();
 
       await act(async () => {
-        return checkoutOperation(sampleAddress);
+        return checkoutOperation(sampleShippingMethod);
       });
 
       expect(isLoading).toBeTruthy();
-      expect(props.values.billingAddress).toEqual(sampleAddress);
+      expect(props.values.shippingMethod).toEqual(sampleShippingMethod);
       expect(props.isLoading).toBeFalsy();
     });
   });
@@ -121,8 +115,8 @@ describe('SetBillingAddress', () => {
     beforeEach(() => {
       client = createApolloClient({
         Mutation: {
-          setBillingAddress: () => {
-            throw new Error('setBillingAddress error');
+          setShippingMethod: () => {
+            throw new Error('setShipping error');
           }
         }
       });
@@ -135,9 +129,9 @@ describe('SetBillingAddress', () => {
       }
     });
 
-    it('should not set billing address on Checkout context', async () => {
-      let checkoutOperation: CheckoutOperationFunction<SetBillingAddressResponse, CheckoutAddress>;
-      let checkoutOperationResult: MutationResult<SetBillingAddressResponse>;
+    it('should not set shipping method on Checkout context', async () => {
+      let checkoutOperation: CheckoutOperationFunction<SetShippingMethodResponse, ShippingMethod>;
+      let checkoutOperationResult: MutationResult<SetShippingMethodResponse>;
       let props: CheckoutRenderProps;
       let isLoading: boolean;
       wrapper = mount(
@@ -151,14 +145,14 @@ describe('SetBillingAddress', () => {
                 }
 
                 return (
-                  <SetBillingAddress>
+                  <SetShippingMethod>
                     {(operation, result) => {
                       checkoutOperation = operation;
                       checkoutOperationResult = result;
 
                       return <div />;
                     }}
-                  </SetBillingAddress>
+                  </SetShippingMethod>
                 );
               }}
             </Checkout>
@@ -167,14 +161,14 @@ describe('SetBillingAddress', () => {
       );
 
       expect(props.isLoading).toBeFalsy();
-      expect(props.values.billingAddress).toBeUndefined();
+      expect(props.values.shippingMethod).toBeUndefined();
 
       await act(async () => {
-        return checkoutOperation(sampleAddress);
+        return checkoutOperation(sampleShippingMethod);
       });
 
       expect(isLoading).toBeTruthy();
-      expect(props.values.billingAddress).toBeUndefined();
+      expect(props.values.shippingMethod).toBeUndefined();
       expect(checkoutOperationResult.error).toBeDefined();
       expect(props.isLoading).toBeFalsy();
     });
